+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/gfx/enums/FileFormatType.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library.
  */
-package com.jpexs.decompiler.flash.tags.gfx;
+package com.jpexs.decompiler.flash.tags.gfx.enums;
 
 import java.lang.reflect.Field;